second commit with map and extoernal json data

+++ src/components/Expenses/ExpenseDate.js
@@ -1,10 +1,9 @@
 import './ExpenseDate.css';
 
 export default function ExpenseDate(props) {
-  const month = props.date.toLocaleString('en-US', { month: 'long' });
+  const month = props.date.toLocaleString('en-US', { month: 'short' });
   const day = props.date.toLocaleString('en-US', { day: '2-digit' });
   const year = props.date.getFullYear();
-
   return (
     <div className="expense-date">
       <div className="expense-date__month">{month}</div>
